add hbs store and move hbs functionality from holo store
